Replace save and load functions with fsc.hdf5-io
