+++ .gitignore
@@ -52,3 +52,6 @@ playground
 
 # mypy
 .mypy_cache
+
+# Atom ide-python
+.ropeproject
