+++ tbmodels/_tb_model.py
@@ -1,7 +1,6 @@
 import re
 import os
 import copy
-import json
 import time
 import warnings
 import itertools
