+++ tests/test_hdf5.py
@@ -24,3 +24,18 @@ def test_hdf5_consistency_file(get_model, models_equal, kwargs):
         model1.to_hdf5_file(tmpf.name)
         model2 = tbmodels.Model.from_hdf5_file(tmpf.name)
     models_equal(model1, model2)
+
+
+@pytest.fixture(params=['InAs_nosym.hdf5'])
+def hdf5_sample(sample, request):
+    return sample(request.param)
+
+
+def test_hdf5_load_freefunc(hdf5_sample):
+    res = tbmodels.io.load(hdf5_sample)
+    assert isinstance(res, tbmodels.Model)
+
+
+def test_hdf5_load_method(hdf5_sample):
+    res = tbmodels.Model.from_hdf5_file(hdf5_sample)
+    assert isinstance(res, tbmodels.Model)
